+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1201645038896279 - Add date access methods to LDLFileInterface
 - feature/1201645038896283 - Add sort by date methods to FileTreeInterface
 - feature/1201645038896275 - Add file ownership capabilities to LDLFileInterface
+- feature/1201663881407974 - Add FileTreeInterface::getFileByName(string $name) : ?LDLFileInterface
 
 ### Changed